update: loop invariant example

+++ examples/loop_invariant.m
@@ -7,15 +7,15 @@ setup()
 
 mainloop()
 {
-    int8 i = 1; // For outer loop
+    uint32 i = 1; // For outer loop
 
     while (i <= 10)
     {
-        int8 L = i * (n + 2);
-        int8 j = 1; // For inner loop
+        uint16 L = i * (n + 2);
+        uint32 j = 1; // For inner loop
         while (j <= 10)
         { 
-            int8 R = L * 10 * j; // Interesting calculation
+            uint16 R = L * 10 * j; // Interesting calculation
             j = j + 1; // Increment inner loop
         }
         i = i + 1; // Increment outer loop
